pyproject.toml
[project]
name = "unicon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0.32",
    "python-dotenv",
    "pydantic>=2.9.1",
    "fastapi[standard]>=0.114.1",
    "pyjwt>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "requests>=2.32.3",
    "rich>=13.8.1",
    "pika>=1.3.2",
    "alembic>=1.13.3",
    "alembic-postgresql-enum>=1.3.0",
    "psycopg2-binary>=2.9.9",
    "sqlmodel>=0.0.22",
    "libcst>=1.5.1",
    "permify>=1.0.2",
    "minio>=7.2.15",
    "pathvalidate>=3.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "mypy>=1.11.2",
    "types-passlib>=1.7.7.20240819",
    "types-requests>=2.32.0.20240914",
    "pre-commit>=3.8.0",
    "types-pika-ts>=1.3.0.20241105",
    "ruff>=0.7.2",
]

[tool.ruff]
line-length = 100
lint.select = ["I", "F", "UP", "B", "SIM", "TCH"]

[tool.mypy]
check_untyped_defs = true

[dependency-groups]
cli = [
    "typer>=0.15.1",
]
